Fix: Better error message

+++ app/api/article/route.ts
@@ -40,7 +40,11 @@ export async function POST(req: Request) {
   let article = formToObject<B_NewArticle>(form_data);
 
   if ( !isAnImage(article.image) || typeof article.image === "string" ) {
-    return httpResponse(StatusCode.Unauthorized);
+    return httpResponse(
+      StatusCode.Unauthorized, 
+      undefined, 
+      "Veuillez mettre une image approprier afin de cree un article"
+    );
   }
 
   if ( !article.image ) return httpResponse(StatusCode.Unauthorized);
@@ -56,27 +60,51 @@ export async function POST(req: Request) {
   if ( !image_exist ) {
      const new_image = await Image.new({ path: article.image, id_used: [] })
 
-    if ( new_image !== StatusCode.Success ) return httpResponse(new_image)
+    if ( new_image !== StatusCode.Success ) {
+      return httpResponse(
+        new_image, 
+        undefined, 
+        "Il y a eu une erreur lors de l'enregistrement de votre image, veuillez reessayer."
+      )
+    }
 
     const github_manager = new Github();
     const pushed_pass = await github_manager.push_gihtub_files([image]);
 
-    if ( !pushed_pass ) return httpResponse(StatusCode.InternalError);
+    if ( !pushed_pass ) {
+      return httpResponse(
+        StatusCode.InternalError, 
+        undefined, 
+        "Il y a eu une erreur lors de la sauvegarde de votre image, veuillez reessayer."
+      );
+    }
   }
 
   const new_article = await Article.new(article);
 
-  if ( new_article.status !== StatusCode.Success || image_exist ) return httpResponse(new_article.status)
+  if ( new_article.status !== StatusCode.Success ){ 
+    return httpResponse(
+      new_article.status, 
+      undefined, 
+      new_article.message
+    )
+  }
 
   const image_db = await Image.get(article.image);
 
-  if ( !image_db ) return httpResponse(StatusCode.NotFound) 
+  if ( !image_db ) {
+    return httpResponse(
+      StatusCode.NotFound, 
+      undefined, 
+      "Une erreur est survenu, lors de l'assignement de votre image a votre article, veuillez reessayer"
+    )
+  }
 
-  image_db.id_used = [...image_db.id_used, new_article._id]
+  image_db.id_used.push(new_article._id)
 
   const update_image = await Image.update(image_db)
 
-  return httpResponse(new_article.status);
+  return httpResponse(new_article.status, undefined, new_article.message);
 }
 
 
@@ -88,52 +116,62 @@ export async function PATCH(req: Request) {
   let update_article = formToObject<B_Article>(form_data);
 
   if ( !isAnImage(update_article.image) ) {
-    return httpResponse(StatusCode.Unauthorized);
+    return httpResponse(
+      StatusCode.Unauthorized, 
+      undefined, 
+      "Veuillez mettre une image approprier afin de cree un article"
+    );
   }
 
   let article = await Article.get(update_article._id);
 
   if ( !article ) {
-    return httpResponse(StatusCode.NotFound);
+    return httpResponse(
+      StatusCode.NotFound, 
+      undefined, 
+      "Une erreur est survenu lors de la recuperation de votre article dans le but de le modifier, veuillez reessayer"
   }
+
   
   if ( typeof update_article.image !== "string" && `/images/${update_article.image.name}` !== article.image ) {
+    const github_manager = new Github();
+
     const image: FileType = {
       path: "public/images/" + update_article.image.name,
       content: await toBase64(update_article.image)
     };
     update_article.image = "/images/" + update_article.image.name;
 
-    const delete_if_needed = await Image.delete_id_needed(article.image, update_article._id);
+    const delete_if_needed = await Image.delete_if_needed(article.image, article._id.toString());
 
     if ( delete_if_needed ) {
       let image_name = article.image.split("/")[2];
-
-      const github_manager = new Github();
-      const delete_file = await github_manager.delete_github_file(image_name)
-
-      if ( !delete_file ) {
-        console.log("GITHUB ERROR: Error deleting file")
-        return httpResponse(StatusCode.InternalError);
-      }
+      await github_manager.delete_github_file(image_name)
     }
-
 
     const image_exist = await Image.exist(update_article.image);
 
     if ( !image_exist) {
-      const new_image = await Image.new({ path: update_article.image, id_used: [update_article.id] })
+      const new_image = await Image.new({ path: update_article.image, id_used: [article.id.toString()] })
 
-      if ( new_image !== StatusCode.Success ) return httpResponse(new_image)
+      if ( new_image !== StatusCode.Success ) {
+        return httpResponse(
+          new_image, 
+          undefined, 
+          "Il y a eu une erreur lors de l'enregistrement de votre image, veuillez reessayer."
+        )
+      }
 
-        const pushed_pass = await github_manager.push_gihtub_files([image]);
+      const pushed_pass = await github_manager.push_gihtub_files([image]);
 
       if ( !pushed_pass ) {
         console.log("GITHUB ERROR: Error pushing file")
-        return httpResponse(StatusCode.InternalError);
+        return httpResponse(
+          StatusCode.InternalError, 
+          undefined, 
+          "Il y a eu une erreur lors de la sauvegarde de votre image, veuillez reessayer."
+        );
       }
-    }
-
   } else if ( 
     typeof update_article.image !== "string" && 
     `/images/${update_article.image.name}` === article.image 
@@ -142,7 +180,8 @@ export async function PATCH(req: Request) {
   }
   
   const updating_article = await Article.update(update_article);
-  return httpResponse(updating_article);
+
+  return httpResponse(updating_article.status, undefined, updating_article.message);
 }
 
 export async function DELETE(req: Request) {
@@ -158,15 +197,23 @@ export async function DELETE(req: Request) {
   let article = await Article.get(id);
 
   if ( !article ) {
-    return httpResponse(StatusCode.NotFound);
+    return httpResponse(
+      StatusCode.NotFound, 
+      undefined, 
+      "Une erreur est arriver lors de la determination de l'article a supprimer, veuillez reessayer."
+    );
   }
 
-  let image_name = article.image.split("/")[2];
+  const delete_if_needed = await Image.delete_if_needed(article.image, article._id.toString());
 
-  const github_manager = new Github();
-  const delete_file = await github_manager.delete_github_file(image_name)
+  if ( delete_if_needed ) {
+    let image_name = article.image.split("/")[2];
+
+    const github_manager = new Github();
+    const delete_file = await github_manager.delete_github_file(image_name)
+  }
 
   const delete_article = await Article.delete(id);
 
-  return httpResponse(delete_article);
+  return httpResponse(delete_article.status, undefined, delete_article.message);
 }
